+++ squeal-postgresql/src/Squeal/PostgreSQL.hs
@@ -194,20 +194,26 @@ in
 [User {userName = "Alice", userEmail = Just "alice@gmail.com"},User {userName = "Bob", userEmail = Nothing},User {userName = "Carole", userEmail = Just "carole@hotmail.com"}]
 -}
 module Squeal.PostgreSQL
-  ( module Squeal.PostgreSQL.Binary
+  ( module Squeal.PostgreSQL.Alias
+  , module Squeal.PostgreSQL.Binary
   , module Squeal.PostgreSQL.Definition
   , module Squeal.PostgreSQL.Expression
+  , module Squeal.PostgreSQL.List
   , module Squeal.PostgreSQL.Manipulation
+  , module Squeal.PostgreSQL.PG
   , module Squeal.PostgreSQL.PQ
   , module Squeal.PostgreSQL.Query
   , module Squeal.PostgreSQL.Schema
   , module Squeal.PostgreSQL.Transaction
   ) where
 
+import Squeal.PostgreSQL.Alias
 import Squeal.PostgreSQL.Binary
 import Squeal.PostgreSQL.Definition
 import Squeal.PostgreSQL.Expression
+import Squeal.PostgreSQL.List
 import Squeal.PostgreSQL.Manipulation
+import Squeal.PostgreSQL.PG
 import Squeal.PostgreSQL.PQ
 import Squeal.PostgreSQL.Query
 import Squeal.PostgreSQL.Schema
