+++ squeal-postgresql/src/Squeal/PostgreSQL/Manipulation.hs
@@ -62,8 +62,10 @@ import GHC.TypeLits (KnownSymbol)
 import qualified Generics.SOP as SOP
 import qualified GHC.Generics as GHC
 
-import Squeal.PostgreSQL.Binary
+import Squeal.PostgreSQL.Alias
 import Squeal.PostgreSQL.Expression
+import Squeal.PostgreSQL.List
+import Squeal.PostgreSQL.PG
 import Squeal.PostgreSQL.Render
 import Squeal.PostgreSQL.Query
 import Squeal.PostgreSQL.Schema
