+++ squeal-postgresql/src/Squeal/PostgreSQL/Migration.hs
@@ -147,22 +147,23 @@ module Squeal.PostgreSQL.Migration
   , defaultMain
   ) where
 
-import           Control.Category
-import           Control.Monad
-import           Data.ByteString             (ByteString)
-import           Data.Foldable               (traverse_)
-import           Data.Function               ((&))
-import           Data.List                   ((\\))
-import           Data.Text                   (Text)
-import qualified Data.Text.IO                as T (putStrLn)
-import           Data.Time                   (UTCTime)
-import           Generics.SOP                (K (..))
-import qualified Generics.SOP                as SOP
-import qualified GHC.Generics                as GHC
-import           Prelude hiding ((.), id)
-import           Squeal.PostgreSQL
-import           System.Environment
-import           UnliftIO                    (MonadIO (..))
+import Control.Category
+import Control.Monad
+import Data.ByteString (ByteString)
+import Data.Foldable (traverse_)
+import Data.Function ((&))
+import Data.List ((\\))
+import Data.Text (Text)
+import Data.Time (UTCTime)
+import Generics.SOP (K (..))
+import Prelude hiding ((.), id)
+import Squeal.PostgreSQL
+import System.Environment
+import UnliftIO (MonadIO (..))
+
+import qualified Data.Text.IO as Text (putStrLn)
+import qualified Generics.SOP as SOP
+import qualified GHC.Generics as GHC
 
 -- | A `Migration` is a named "isomorphism" over a given category.
 -- It should contain an inverse pair of `up` and `down`
@@ -437,17 +438,17 @@ defaultMain connectTo migrations = do
       fmap migrationName <$> (unsafePQ (define createMigrations & pqThen (runQuery selectMigrations)) >>= getRows)
 
     displayListOfNames :: [Text] -> IO ()
-    displayListOfNames [] = T.putStrLn "  None"
+    displayListOfNames [] = Text.putStrLn "  None"
     displayListOfNames xs =
-      let singleName n = T.putStrLn $ "  - " <> n
+      let singleName n = Text.putStrLn $ "  - " <> n
       in traverse_ singleName xs
 
     displayUnrunned :: [Text] -> IO ()
     displayUnrunned unrunned =
-      T.putStrLn "Migrations left to run:"
+      Text.putStrLn "Migrations left to run:"
       >> displayListOfNames unrunned
 
     displayRunned :: [Text] -> IO ()
     displayRunned runned =
-      T.putStrLn "Migrations already run:"
+      Text.putStrLn "Migrations already run:"
       >> displayListOfNames runned
